[Test] Fix GCP `testWriteFileMultipleOfChunkSize`
